requirements.txt
pandas
pyyaml
scikit-learn
fastapi
uvicorn
requests
